+++ Sources/Client/Model/Channel/ChannelQuery.swift
@@ -109,8 +109,6 @@ public struct ChannelUpdate: Encodable {
         
         func encode(to encoder: Encoder) throws {
             var container = encoder.container(keyedBy: Channel.EncodingKeys.self)
-            try container.encode(channel.name, forKey: .name)
-            try container.encodeIfPresent(channel.imageURL, forKey: .imageURL)
             channel.extraData?.encodeSafely(to: encoder, logMessage: "📦 when encoding a channel extra data")
         }
     }